+++ src/leveldb/util/env_win.cc
@@ -8,6 +8,7 @@
 #include <map>
 
 
+#include <stdint.h>
 #include "leveldb/env.h"
 
 #include "port/port.h"
@@ -719,7 +720,7 @@ void Win32Logger::Logv( const char* format, va_list ap )
             int(st.wMinute),
             int(st.wMinute),
             int(st.wMilliseconds),
-            static_cast<long long unsigned int>(thread_id));
+            static_cast<uint64_t>(thread_id));
 
         // Print the message
         if (p < limit) {
