+++ src/leveldb/util/logging.cc
@@ -4,6 +4,7 @@
 
 #include "util/logging.h"
 
+#include <stdint.h>
 #include <errno.h>
 #include <stdarg.h>
 #include <stdio.h>
@@ -15,7 +16,7 @@ namespace leveldb {
 
 void AppendNumberTo(std::string* str, uint64_t num) {
   char buf[30];
-  snprintf(buf, sizeof(buf), "%" PRIu64, (unsigned long long) num);
+  snprintf(buf, sizeof(buf), "%" PRIu64, num);
   str->append(buf);
 }
